Moved FeedImageMapper class to its own file.
